+++ verl/utils/vllm/utils.py
@@ -126,3 +126,121 @@ class VLLMHijack:
 def is_version_ge(pkg: str = "vllm", minver: str = "0.7.3"):
     """check if the package version is greater than or equal to the minimum version"""
     return vs.parse(get_version(pkg)) >= vs.parse(minver)
+
+
+# Whether the installed vLLM ships ``BaseLayerWithLoRA.load_weights`` (landed in
+# #39935 / ``b50fdebce0``). On newer vLLM the LoRA layer's ``load_weights``
+# delegates to ``AutoWeightsLoader(base_layer)`` and expects inner names without
+# ``.base_layer.``; on released vLLM (False) the model's ``load_weights``
+# recurses into the ``base_layer`` child and matches the suffixed name.
+_HAS_LORA_LOAD_WEIGHTS = False
+
+try:
+    from vllm.lora.layers.base import BaseLayerWithLoRA
+
+    _HAS_LORA_LOAD_WEIGHTS = "load_weights" in BaseLayerWithLoRA.__dict__
+except ImportError:
+    pass
+
+
+def resolve_weight_name(model, name: str, model_weight_names: set[str]) -> str:
+    """Reconcile an incoming weight-sync name with the live vLLM namespace.
+
+    Toggles one ``.base_layer`` segment (strip it for non-LoRA-wrapped modules
+    on all vLLM versions and for LoRA-wrapped modules on newer vLLM; add it for
+    LoRA-wrapped leaves on released vLLM) and routes packed routed-expert
+    aliases under the LoRA ``base_layer`` so ``AutoWeightsLoader`` reaches
+    ``MoERunner.load_weights``.
+
+    The mapper / ``packed_modules_mapping`` are consulted only to *decide* the
+    toggle; the return is always the unmapped name (vLLM's ``load_weights`` does
+    the stacking -- returning a mapped name would double-map, e.g.
+    ``in_proj_qkvz`` -> ``in_proj_qkvzz``, corrupting shard_id).
+    """
+    mapper = getattr(model, "hf_to_vllm_mapper", None)
+    packed = getattr(model, "packed_modules_mapping", None) or {}
+    leaf = name.rsplit(".", 1)[-1]
+    is_leaf = leaf in {"weight", "bias"} or leaf.endswith(("_weight", "_bias"))
+
+    def _exists(candidate: str) -> bool:
+        if candidate in model_weight_names:
+            return True
+        if mapper is not None:
+            mapped = mapper.apply_list([candidate])
+            mapped = mapped[0] if mapped else candidate
+            if mapped != candidate and mapped in model_weight_names:
+                return True
+        # Packed-owner lookup (q/k/v -> qkv) via packed_modules_mapping.
+        # Hoisted out of the mapper guard so models with
+        # packed_modules_mapping but no hf_to_vllm_mapper (e.g. Llama) also
+        # reach it.
+        if packed and "." in candidate:
+            parts = candidate.split(".")
+            mi = -3 if len(parts) >= 3 and parts[-2] == "base_layer" else -2
+            if -mi <= len(parts):
+                # packed is {owner: [unpacked,...]}; invert to {unpacked: owner}.
+                rev = {u: p for p, us in packed.items() for u in us}
+                owner = rev.get(parts[mi])
+                for pn in (owner,) if owner else ():
+                    pp = parts.copy()
+                    pp[mi] = pn
+                    joined = ".".join(pp)
+                    if joined in model_weight_names:
+                        return True
+                    if mapper is not None:
+                        mp = mapper.apply_list([joined])
+                        mp = mp[0] if mp else joined
+                        if mp != candidate and mp in model_weight_names:
+                            return True
+        return False
+
+    # Strip ``.base_layer.``:
+    # - On models with NO ``.base_layer.`` params at all (auxiliary models like
+    #   the MTP drafter, or non-LoRA-wrapped modules such as the Qwen3.5-VL vision
+    #   merger): strip unconditionally -- there is no ``base_layer`` child to
+    #   recurse into on any vLLM version, so the inner name is the only one that
+    #   could match.
+    # - On LoRA-wrapped models (which DO have ``.base_layer.`` params): strip only
+    #   on newer vLLM, where ``BaseLayerWithLoRA.load_weights`` delegates to
+    #   ``AutoWeightsLoader(base_layer)`` expecting inner names. On released vLLM
+    #   keep the suffix (the loader recurses into ``base_layer`` and matches it)
+    #   and fall through to the leaf-add path below.
+    if ".base_layer." in name:
+        model_has_base_layer = any(".base_layer." in n for n in model_weight_names)
+        if not model_has_base_layer or _HAS_LORA_LOAD_WEIGHTS:
+            return name.replace(".base_layer.", ".", 1)
+
+    if _exists(name):
+        return name
+
+    # Packed routed-expert alias: route under the LoRA base_layer so
+    # AutoWeightsLoader reaches MoERunner.load_weights (FusedMoE3DWithLoRA has no
+    # load_weights itself). Only for non-leaf, non-routed names.
+    marker = ".mlp.experts."
+    idx = name.find(marker)
+    if idx != -1:
+        tail = name[idx + len(marker) :]
+        if (
+            tail
+            and ".base_layer." not in tail
+            and not is_leaf
+            and any("mlp.experts.base_layer." in n for n, _ in model.named_parameters(remove_duplicate=False))
+        ):
+            return name.replace(marker, marker + "base_layer.", 1)
+
+    if ".mlp.experts.base_layer." in name and not is_leaf:
+        stripped = name.replace(".base_layer.", ".", 1)
+        if _exists(stripped):
+            return stripped
+
+    # Leaf weight/bias: on vLLM versions without BaseLayerWithLoRA.load_weights,
+    # add ``.base_layer.`` so AutoWeightsLoader recurses into the base_layer
+    # child; on newer vLLM, the strip above already handled it.
+    if is_leaf:
+        if not _HAS_LORA_LOAD_WEIGHTS and ".base_layer." not in name:
+            prefix, lf = name.rsplit(".", 1)
+            alt = f"{prefix}.base_layer.{lf}"
+            if alt != name and _exists(alt):
+                return alt
+
+    return name
